tests: Remove references to Project Gutenberg
The test file for the "lines" test is the Project Gutenberg edition of
"Alice in Wonderland" (taken from the DMD git repository, where it is
located at test/runnable/extra-files/alice30.txt). Although the book
is now in the public domain, Project Gutenberg imposes additional
restrictions upon redistribution for the file in its original form.

The "small print" section in the text file says:

> DISTRIBUTION UNDER "PROJECT GUTENBERG-tm"
> You may distribute copies of this etext electronically, or by
> disk, book or any other medium if you either delete this
> "Small Print!" and all other references to Project Gutenberg,
> or:
> [...]

This commit removes the the small print section at the beginning, thus
complying with the redistribution requirements.

Fixes #50.