Don't move code marked as no-remove
If the user is marking some code as not to be removed, they likely
intend that it should not be moved around either. But, even if this
assumption is false, the effort required to move (concatenate) code
manually is much smaller than having to deal with a false reduction
due to code having been moved around.

Fixes #61.