dustmite: Implement non-linear lookahead prediction
Previously, the prediction/lookahead was completely linear: one result
was predicted (a negative), and further lookahead jobs were started
assuming that result.

This was suboptimal.

Change the prediction function to return probability instead of a
boolean, then build a tree of probabilities. Keep expanding the tree
at the most likely leaf.