Remove the extension stripping for single-file input
As extensions are not an intrinsic aspect of filesystems, but are part
of user-defined semantics for how to interpret file names, don't
hard-code file name extension manipulation into the program.

This collisions with input file names such as
"src.result" (in which case dustmite tried to create "src.result" by
appending its own ".result" suffix to the extension-stripped file name
of the input file).