dustmite: Handle errors
A reduction might create a tree which is outright impossible to write
- for example, a symlink with a zero-length target. If we ever allow
making the oracle (tester) a file, spawnProcess may also cause an
exception within DustMite (as opposed to just returning a non-zero
process exit code).

Add code to handle errors in both the single-thread and lookahead
paths. These are categorized as their own source.